requirements.txt
numpy==1.21.4
tensorflow==2.7
opencv-contrib-python==4.5.4
imutils==0.5.4
matplotlib==3.5.1
patchify
focal-loss==0.0.7
keras==2.7.0
pandas==1.3.5
pillow==8.4.0
sklearn==1.0.1
tifffile==2021.11.2
split-folder==0.4.3